+++ ouzel/audio/empty/EmptyAudioDevice.hpp
@@ -16,9 +16,9 @@ namespace ouzel
             public:
                 AudioDevice(uint32_t initBufferSize,
                             uint32_t initSampleRate,
-                            uint16_t initChannels,
+                            uint32_t initChannels,
                             const std::function<void(uint32_t frames,
-                                                     uint16_t channels,
+                                                     uint32_t channels,
                                                      uint32_t sampleRate,
                                                      std::vector<float>& samples)>& initDataGetter);
