+++ ouzel/audio/mixer/Mixer.cpp
@@ -13,7 +13,7 @@ namespace ouzel
         namespace mixer
         {
             Mixer::Mixer(uint32_t initBufferSize,
-                         uint16_t initChannels,
+                         uint32_t initChannels,
                          const std::function<void(const Event&)>& initCallback):
                 bufferSize(initBufferSize),
                 channels(initChannels),
@@ -210,7 +210,7 @@ namespace ouzel
                 }
             }
 
-            void Mixer::getSamples(uint32_t frames, uint16_t channels, uint32_t sampleRate, std::vector<float>& samples)
+            void Mixer::getSamples(uint32_t frames, uint32_t channels, uint32_t sampleRate, std::vector<float>& samples)
             {
                 process();
